Pull tracker utils out of MountingResetTests
