+++ server/core/src/test/java/dev/slimevr/unit/MountingResetTests.kt
@@ -4,14 +4,15 @@ import com.jme3.math.FastMath
 import dev.slimevr.VRServer.Companion.getNextLocalTrackerId
 import dev.slimevr.tracking.trackers.Tracker
 import dev.slimevr.tracking.trackers.udp.IMUType
+import dev.slimevr.unit.TrackerUtils.assertAnglesApproxEqual
+import dev.slimevr.unit.TrackerUtils.deg
+import dev.slimevr.unit.TrackerUtils.yaw
 import io.github.axisangles.ktmath.EulerAngles
 import io.github.axisangles.ktmath.EulerOrder
 import io.github.axisangles.ktmath.Quaternion
-import org.junit.jupiter.api.AssertionFailureBuilder
 import org.junit.jupiter.api.DynamicTest
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestFactory
-import kotlin.math.*
 
 /**
  * Tests [TrackerResetsHandler.resetMounting]
@@ -22,8 +23,8 @@ import kotlin.math.*
 class MountingResetTests {
 
 	@TestFactory
-	fun testResetAndMounting(): List<DynamicTest> = directions.flatMap { e ->
-		directions.map { m ->
+	fun testResetAndMounting(): List<DynamicTest> = TrackerUtils.directions.flatMap { e ->
+		TrackerUtils.directions.map { m ->
 			DynamicTest.dynamicTest(
 				"Full and Mounting Reset Test of Tracker (Expected: ${deg(e)}, reference: ${deg(m)})",
 			) {
@@ -34,7 +35,7 @@ class MountingResetTests {
 
 	private fun checkResetMounting(expected: Quaternion, reference: Quaternion) {
 		// Compute the pitch/roll for the expected mounting
-		val trackerRot = (expected * (frontRot / expected))
+		val trackerRot = (expected * (TrackerUtils.frontRot / expected))
 
 		val tracker = Tracker(
 			null,
@@ -119,7 +120,7 @@ class MountingResetTests {
 		val expected = Quaternion.SLIMEVR.RIGHT
 		val reference = EulerAngles(EulerOrder.YZX, FastMath.PI / 8f, FastMath.HALF_PI, 0f).toQuaternion()
 		// Compute the pitch/roll for the expected mounting
-		val trackerRot = (expected * (frontRot / expected))
+		val trackerRot = (expected * (TrackerUtils.frontRot / expected))
 
 		val tracker = Tracker(
 			null,
@@ -158,52 +159,5 @@ class MountingResetTests {
 			resultYaw2,
 			"Resulting rotation after yaw reset is not equal to reference yaw (${deg(expectedYaw2)} vs ${deg(resultYaw2)})",
 		)
-	}
-
-	/**
-	 * Makes a radian angle positive
-	 */
-	private fun posRad(rot: Float): Float {
-		// Reduce the rotation to the smallest form
-		val redRot = rot % FastMath.TWO_PI
-		return abs(if (rot < 0f) FastMath.TWO_PI + redRot else redRot)
-	}
-
-	/**
-	 * Gets the yaw of a rotation in radians
-	 */
-	private fun yaw(rot: Quaternion): Float = posRad(rot.toEulerAngles(EulerOrder.YZX).y)
-
-	/**
-	 * Converts radians to degrees
-	 */
-	private fun deg(rot: Float): Float = rot * FastMath.RAD_TO_DEG
-
-	private fun deg(rot: Quaternion): Float = deg(yaw(rot))
-
-	private fun anglesApproxEqual(a: Float, b: Float): Boolean = FastMath.isApproxEqual(a, b) ||
-		FastMath.isApproxEqual(a - FastMath.TWO_PI, b) ||
-		FastMath.isApproxEqual(a, b - FastMath.TWO_PI)
-
-	private fun assertAnglesApproxEqual(expected: Float, actual: Float, message: String?) {
-		if (!anglesApproxEqual(expected, actual)) {
-			AssertionFailureBuilder.assertionFailure().message(message)
-				.expected(expected).actual(actual).buildAndThrow()
-		}
-	}
-
-	companion object {
-		val directions = arrayOf(
-			Quaternion.SLIMEVR.FRONT,
-			Quaternion.SLIMEVR.FRONT_LEFT,
-			Quaternion.SLIMEVR.LEFT,
-			Quaternion.SLIMEVR.BACK_LEFT,
-			Quaternion.SLIMEVR.FRONT_RIGHT,
-			Quaternion.SLIMEVR.RIGHT,
-			Quaternion.SLIMEVR.BACK_RIGHT,
-			Quaternion.SLIMEVR.BACK,
-		)
-
-		val frontRot = EulerAngles(EulerOrder.YZX, FastMath.HALF_PI, 0f, 0f).toQuaternion()
 	}
 }
